+++ test/test_helpers/email_alert_api_test.rb
@@ -21,4 +21,18 @@ describe GdsApi::TestHelpers::EmailAlertApi do
       assert_email_alert_api_content_change_created("foo" => "bar")
     end
   end
+
+  describe "#assert_email_alert_api_message_created" do
+    before { stub_any_email_alert_api_call }
+
+    it "matches a post request with any empty attributes by default" do
+      email_alert_api.create_message("foo" => "bar")
+      assert_email_alert_api_message_created
+    end
+
+    it "matches a post request subset of attributes" do
+      email_alert_api.create_message("foo" => "bar", "baz" => "qux")
+      assert_email_alert_api_message_created("foo" => "bar")
+    end
+  end
 end